Refactor visualize script; Fixes

+++ visualize/visualize.py
@@ -11,6 +11,12 @@ FOLDER_NAME_IN_TAR = "nperf-results-test" # Normally: "results"
 LOG_FILE = "plots_checklist.md"
 MAPPINGS_FOLDER_PATH = "visualize"
 
+MAPPINGS = {
+    "special": "configs_mapping_special.json",
+    "syscalls": "configs_mapping_syscalls.json",
+    "uring": "configs_mapping_uring.json"
+}
+
 logging.basicConfig(level=logging.INFO , format='%(asctime)s - %(levelname)s - %(message)s')
 
 def create_plots(results_folder: str, csv_folder: str, configs_mapping: dict[str, dict[str, str]]):
@@ -20,7 +26,7 @@ def create_plots(results_folder: str, csv_folder: str, configs_mapping: dict[str
         logging.info(f"Create plot for {config_name}")
         logging.debug(f"Plot config for {config_name}: {plot_config}")
 
-        graph_type = plot_config.get("graph_type", "area")
+        graph_type = plot_config.get("type", "area")
         title = plot_config.get("title")
         x_label = plot_config.get("x_label", "amount_threads") 
         y_label = plot_config.get("y_label", "data_rate_gbit")
@@ -41,68 +47,13 @@ def create_plots(results_folder: str, csv_folder: str, configs_mapping: dict[str
         logging.debug(f"Running command: {command}")
         subprocess.run(command, check=True)
 
+def visualize(folder_name: str, results_folder: str):
+    csv_folder_server = os.path.join(folder_name, f"nperf-server")
+    csv_folder_client = os.path.join(folder_name, f"nperf-client")
 
-def main():
-    logging.info('Starting main function')
-    parser = argparse.ArgumentParser(description="Visualize the results of run.py")
-
-    parser.add_argument("path_to_tar", type=str, help="The relative path to the tar file")
-    parser.add_argument("client_name", type=str, help="The ssh name of the client")
-    parser.add_argument("server_name", type=str, help="The ssh name of the server")
-    parser.add_argument("--results-folder", type=str, default=RESULTS_DIR, help="The folder where the results are stored")
-    parser.add_argument("--folder-name-in-tar", type=str, default=FOLDER_NAME_IN_TAR, help="The folder name in the tar file")
-    args = parser.parse_args()
-
-    logging.info(f"Path to tar file: {args.path_to_tar}")
-    logging.info(f"Client name: {args.client_name}")
-    logging.info(f"Server name: {args.server_name}")
-    temp_folder = args.folder_name_in_tar
-
-    # Create the results folder
-    os.makedirs(args.results_folder, exist_ok=True)
-
-    # Untar the results tar in the current directory
-    with tarfile.open(args.path_to_tar, "r") as tar:
-        tar.extractall(temp_folder)
-
-    logging.info(f"Untar the file {args.client_name}-results.tar.gz inside {temp_folder} into folder nperf-client")
-    client_tar_file = os.path.join(temp_folder, f"{args.client_name}-results.tar.gz")
-    try:
-        with tarfile.open(client_tar_file, "r") as tar:
-            tar.extractall(temp_folder)
-        client_results_folder = os.path.join(temp_folder, "nperf-client")
-        # TODO: Change this to the variable FOLDER_NAME_IN_TAR
-        os.rename(os.path.join(temp_folder, "results"), client_results_folder)
-    except tarfile.TarError as e:
-        logging.error(f"Error extracting tar file: {e}")
-    except FileNotFoundError as e:
-        logging.error(f"File not found: {e}")
-
-    logging.info(f"Untar the file {args.server_name}-results.tar.gz inside {temp_folder} into folder nperf-server")
-    server_tar_file = os.path.join(temp_folder, f"{args.server_name}-results.tar.gz")
-    try:
-        with tarfile.open(server_tar_file, "r") as tar:
-            tar.extractall(temp_folder)
-        server_results_folder = os.path.join(temp_folder, "nperf-server")
-        # TODO: Change this to the variable FOLDER_NAME_IN_TAR
-        os.rename(os.path.join(temp_folder, "results"), server_results_folder)
-    except tarfile.TarError as e:
-        logging.error(f"Error extracting tar file: {e}")
-    except FileNotFoundError as e:
-        logging.error(f"File not found: {e}")
-        
-    mappings = {
-        "special": "configs_mapping_special.json",
-        "syscalls": "configs_mapping_syscalls.json",
-        "uring": "configs_mapping_uring.json"
-    }
-
-    csv_folder_server = os.path.join(temp_folder, f"nperf-server")
-    csv_folder_client = os.path.join(temp_folder, f"nperf-client")
-
-    for key, filename in mappings.items():
+    for key, filename in MAPPINGS.items():
         file_path = os.path.join(MAPPINGS_FOLDER_PATH, filename)
-        results_folder_path = os.path.join(args.results_folder, key)
+        results_folder_path = os.path.join(results_folder, key)
 
         try:
             os.makedirs(results_folder_path, exist_ok=True) 
@@ -116,8 +67,72 @@ def main():
         except json.JSONDecodeError as e:
             logging.error(f"Error decoding JSON from {filename}: {e}")
 
-    # Remove the temporary folder
-    shutil.rmtree(temp_folder)
+
+def unpack_tar(tar_path: str, folder_name: str, client_name: str, server_name: str):
+    # Untar the results tar in the current directory
+    with tarfile.open(tar_path, "r") as tar:
+        tar.extractall(folder_name)
+
+    logging.info(f"Untar the file {client_name}-results.tar.gz inside {folder_name} into folder nperf-client")
+    client_tar_file = os.path.join(folder_name, f"{client_name}-results.tar.gz")
+
+    try:
+        with tarfile.open(client_tar_file, "r") as tar:
+            tar.extractall(folder_name)
+        client_results_folder = os.path.join(folder_name, "nperf-client")
+        # TODO: Change this to the variable FOLDER_NAME_IN_TAR
+        os.rename(os.path.join(folder_name, "results"), client_results_folder)
+    except tarfile.TarError as e:
+        logging.error(f"Error extracting tar file: {e}")
+    except FileNotFoundError as e:
+        logging.error(f"File not found: {e}")
+
+    logging.info(f"Untar the file {server_name}-results.tar.gz inside {folder_name} into folder nperf-server")
+    server_tar_file = os.path.join(folder_name, f"{server_name}-results.tar.gz")
+
+    try:
+        with tarfile.open(server_tar_file, "r") as tar:
+            tar.extractall(folder_name)
+        server_results_folder = os.path.join(folder_name, "nperf-server")
+        # TODO: Change this to the variable FOLDER_NAME_IN_TAR
+        os.rename(os.path.join(folder_name, "results"), server_results_folder)
+    except tarfile.TarError as e:
+        logging.error(f"Error extracting tar file: {e}")
+    except FileNotFoundError as e:
+        logging.error(f"File not found: {e}")
+
+
+def main():
+    logging.info('Starting main function')
+    parser = argparse.ArgumentParser(description="Visualize the results of run.py")
+
+    parser.add_argument("path_to_tar", type=str, help="The relative path to the tar file")
+    parser.add_argument("client_name", type=str, help="The ssh name of the client")
+    parser.add_argument("server_name", type=str, help="The ssh name of the server")
+    parser.add_argument("--results-folder", type=str, default=RESULTS_DIR, help="The folder where the results are stored")
+    parser.add_argument("--folder-name-in-tar", type=str, default=FOLDER_NAME_IN_TAR, help="The folder name in the tar file")
+    parser.add_argument("--use-existing", action="store_true", help="Use existing temp folder data instead of extracting the tar file.")
+
+    args = parser.parse_args()
+
+    logging.info(f"Path to tar file: {args.path_to_tar}")
+    logging.info(f"Client name: {args.client_name}")
+    logging.info(f"Server name: {args.server_name}")
+    temp_folder = args.folder_name_in_tar
+
+    # Create the results folder
+    os.makedirs(args.results_folder, exist_ok=True)
+
+    if os.path.exists(temp_folder) and not args.use_existing:
+        logging.warning(f"Removing existing folder {temp_folder}")
+        shutil.rmtree(temp_folder)
+
+    if not args.use_existing:
+        unpack_tar(args.path_to_tar, temp_folder, args.client_name, args.server_name)
+
+    visualize(temp_folder, args.results_folder)
+
+
 
 if __name__ == '__main__':
     logging.info('Starting visualize script')
